Created `profiler` module out of `runner`.
